requirements.txt
pandas
seaborn
torch
torchvision
numpy
scikit-learn
matplotlib
pillow
hdbscan
scikit-image
